improve audio volume of knocks
This is based on the fact that audio volume in dB is

20 log10 (P / P0)

where P is sound power and P0 is the reference power that will be
treated as "volume=0"
In this commit we choose a reference impact velocity to count as "vol=0"
and then we make P = m*impact_vel + b, where m and b are free parameters
to be determined. I chose the parameters by feel.