made "return to menu" on GameBase scroll the background back up
This is working ok, but required some reorganization of scenes (putting boulders and the floor as children of "bg", so that they move with it). Not sure if this is elegant, but seems to make sense to me. But if you don't like how it's done, please feel free to modify it or even just roll back to an earlier commit and proceed as if it didn't happen-- no offense taken!