change HUDZone type to VBoxContainer
Apparently the whole point of a control node is to have its position and
size be managed by its parent, so it makes sense to have the root one
represent the way we want to layout the children.

No visual difference but now if you ever do play with the HUDZone
position or size, then you should see its children automatically arrange
themselves properly.

To get this to work we also have to  set anchors and size flags for the children.