add Logo script that generates logo on title screen
Yes there is a lot of code duplication here D:
Logo.gd copies a lot from Rock.gd, except that generate_polygon() is
modified and a new method "random_point_in_disk_and_outside_disk" is added.

Really there should be a single class extending Node2D which creates a textured
rock polygon, and this should be made a subnode of any Rock (RigidBody2D) object.
This rock polygon class could then be extended to make the logo as well.
I might change things so that this is how it works.