introduce new rock throwing physics based on forces
Cursor position now influences rock by applying a force,
rather than by directly influencing position.
The force pulls the rock towards the cursor with a strength
proportional to the distance between the rock and the cursor.
This would normally cause oscillation, but there is also a damping
that is set to be just enough for each rock to stop oscillations.

This setup represents the following ideas:
- The holder does NOT have enough control to get the rock to where they
want as fast as they want. It will take time to accelerate the rock,
and that depends on the mass.
- The holder DOES have enough control to stop the rock from moving
at the point that they want. For a more massive rock, which would
be harder to stop, the holder has enough judgement to accelerate
the rock to the desired position more gently so that they can still
put it to a stop with zero overshooting.

With these chanegs, heavy rocks and light rocks really do need to be handled
differently. It is no longer possible to make every rock into a bullet,
though it still works for some smaller rocks.