collisions now trigger release of held rock
This happens after having held a rock for a certain time (half a second)
This gives enough time to pick up a rock if it's in a pile of other rocks,
but it will release a rock if you hit an obstacle with it.
This prevents the player from trying to phase a rock through a solid
and having the forces cause the rock to freak out and bounce all over the place.