cap fps to 60
I think this is just going to be for "debug" mode? i.e. our testing?
Will it apply to the exported project?

Anyway, this stops our peuptops from freaking out when we run the project.
I realized I was running at over 2000 fps and that's why the peuptop was dying