don't lock out throwing spawning rocks in chlg mode
Instead let player throw, and also keep retrying to spawn rocks
till one succeeds.

Basically just spawn a rock on the boulder whenever the previous target
rock has been missing for long enough.

And be aware of whether or not a throwing rock has touched the target rock before.
If one has, then when that target rock falls we can count it as a point
(regardless of whether the fall was due to poor initial placement)