challenge mode now starts with a pair of rocks
You get a new rock when one is lost, receiving 10 rocks total. The number of new rocks remaining can be viewed by the player on a label. I'm trying to figure out if any of the code in check_endgame_condition() is now unnecessary, but it doesn't seem to be.