now you can throw rocks that re-enter throwzone
This seems fair because of the throwing area now being smaller, but we can backpedal on this one if you don't like it! I think it would make sense in physical KTR also :D It's done by keeping a separate list up to date that keeps track of the rocks in the throwzone.