change delay rules for throwable rock spawning
This fixes an issue I introduced in 5723ce248ee3f73fb17f18f34ee6192dd0d377ac
The issue is that there was just one delay timer managing the spawning of up
to two throwable rocks. Sometimes the timer gets restarted twice, and then
triggers the spawning of only one rock.

Now the setup is this:
When a throwable rock gets lost, we start a timer to replace it.
If a second throwable rock is lost that needs to be replaced, then
that one is just replaced instantly.