changed how settings config is loaded, fixes #21
Previously the default_settings resource was internal to Main menu scene
Now it is an external resource in a file, default_settings.tres
When we export the project, this file can be packaged into the executable,
and the user never sees it.
We do not export the settings.tres file-- that will be something the user can see.

Launching the game executable with no settings.tres, for example when launching for
the first time, will no longer cause a freakout. Instead it just loads
default settings. A settings.tres only gets saved and used if the user applies
some settings.