enlarge throwzone floor
this reduces the chance of some accidents where player might
pick a rock off throwzone floor, keep it hovering at a low
height above the floor, and then have a newly spawned rock phase
into the same collision layer. this accident can still happen,
but it's rarer now. And the consequence is not that bad:
two rocks inside each other