fix #25
The problem was that we would only retire the old target rock when we spawn a new one,
rather tahn when the old target rock is initially knocked off

During that delay, the old target rock can roll into the throwzone.
Then it ends up getting retired in the throwzone.
This includes setting it's reason for not being holdable,
so it ends up in the throwzone and unholdable.

Holdabiity is used by LineOfPebbles.gd to determine whether or not
to emit rock lost and rock regained signals. This is why the
throwzone rocks count was also gettign messed up