improvement to rock schwoop_delete
(1) Sometimes for some reason rock is so obstructed that it cannot leave the screen.
In these cases it looks bad for it to suddenly pop out of existence after three seconds.
Instead, it now fades out. So you either see the offending rock sucked up out of the screen,
or you see it float gently and fade out.

(2) When schwoop is initiated, we now zero the velocity. Previously if the offending
rock had a downward velocity, then it took a while for the schwooping force to get it off the screen.