shorten throwzone floor
I originally made this taller because I thought it would reduce the likelihood
of problems mentioned in issue #24,
but it didn't help that much and it caused more harm.

(The harm of making it tall has to do with situations where you pick a rock
of the ground to throw and it crosses the path of a just-spawned rock.
If throwzone floor area is tall, you are more likely to get an unwanted collision.)