make justspawned rock unholdable when #24 happens
When the thing in issue #24 happens,
the rock that was just spawned (and that is overlapping
what is probably a recently held rock) will not be holdable.
This allows the player to grab just one of the two overlapping rocks
and resolve the overlap.

Also the rock that can be held will be raised in front visually