make the right wall tall again, and contain spawns
Spawned rocks are now guaranteed to be inside the throw zone completely,
so they don't freak out against the right wall

I've made the right wall tall again because, even though it's now possible
to notice it, it would be much worse if a just-spawned rock managed to
get over the wall in its just-spawned state.

It could overlap another rock and then the overlap could never get resolved
because it's out of the player's control!