prevent infinite loop when spawning rocks
Sometimes when spawning two rocks on the spawn line, the first one
chosen is too large and too poorly positioned to allow for a second one
to be possible. This results in an infinite loop and the game lock up.
What should happen instead is that after a certain number of positioning attempts,
the entire spawn_rocks command is tried again