used method described in #12 to lighten high notes
The treble cut is a really good idea! I see that it is just decreasing the amplitude of higher frequencies so it shouldn't impact the entire song, and it doesn't sound to me like it is. This seems like a good middle ground where the higher notes are still discernible but not piercing. fix #12