address #40
I believe it is fixed, and the problem seems to be gone with no new bugs after testing it a bit, but I wanted to leave the issue open so we can look at something together.

Changes in a bunch of different places were made to accomplish the results here, but these are the main two changes that caused other things to change:
(1) Now, spawn_rocks looks for other rocks not only within the set of other rocks it has spawned so far, but within ALL of RockList.
(2) It turned out this wasn't enough for the zero-G rocks. Sometimes the line was just too cramped for them to find room. So the spawn line is now a spawn BOX, which required topmost_vertex() and botmost_vertex() in RockPolygon.gd and Rock.gd.

One thing I want to look at together (or just me looking at it further) before closing the issue is the new yield statement in GameBase.gd. Is it necessary, is it peupy/problematic?