adjusted justspawned collisionness
Stopped rocks from sleeping when the justspawned collisionness ends. The reason is that it causes an odd behavior in low gravity (the rocks, unsurprisingly, would fall a bit and then stop in their tracks). I think we previously agreed the sleeping was nice but not necessary, so I think just removing it is fine for now.